+++ packages/@atjson/offset-annotations/src/annotations/telegram-embed.ts
@@ -1,6 +1,0 @@
-import { IframeEmbed } from "./iframe-embed";
-
-export class TelegramEmbed extends IframeEmbed {
-  static type = "telegram-embed";
-  static vendorPrefix = "offset";
-}