+++ packages/@atjson/offset-annotations/src/annotations/threads-embed.ts
@@ -1,10 +1,16 @@
 import { BlockAnnotation } from "@atjson/document";
 
-export class TikTokEmbed extends BlockAnnotation<{
+export class ThreadsEmbed extends BlockAnnotation<{
   /**
-   * The URL to the TikTok video on www.tiktok.com
+   * The URL to the Threads post
    */
   url: string;
+
+  /**
+   * Content of the Threads post, which allows to store
+   * text of the post content if that is wanted.
+   */
+  content?: string;
 
   /**
    * Layout information, used to indicate mutually
@@ -16,13 +22,7 @@ export class TikTokEmbed extends BlockAnnotation<{
    * A named identifier used to quickly jump to this item
    */
   anchorName?: string;
-
-  /**
-   * The post content at the time of embedding, ususally
-   * a textual representation of the content with some links.
-   */
-  content?: string;
 }> {
-  static type = "tiktok-embed";
+  static type = "threads-embed";
   static vendorPrefix = "offset";
 }