+++ packages/@atjson/offset-annotations/src/utils/index.ts
@@ -2,10 +2,12 @@ import { getClosestAspectRatio } from "./aspect-ratios";
 import * as SocialURLs from "./social-urls";
 import * as VideoURLs from "./video-urls";
 import { convertHTMLTablesToDataSet } from "./convert-html-tables";
+import { TextAlignment } from "./enums";
 
 export {
   convertHTMLTablesToDataSet,
   getClosestAspectRatio,
   SocialURLs,
+  TextAlignment,
   VideoURLs,
 };